Collaborator removal security

+++ server/src/Application/Command/Board/RemoveCollaboratorCommandHandler.php
@@ -2,6 +2,8 @@
 
 namespace App\Application\Command\Board;
 
+use App\Domain\Exception\Collaborator\NoOwnerLeftException;
+use App\Domain\Model\Collaborator;
 use App\Domain\Repository\CollaboratorRepositoryInterface;
 
 class RemoveCollaboratorCommandHandler
@@ -23,12 +25,28 @@ class RemoveCollaboratorCommandHandler
 
     /**
      * @var RemoveCollaboratorCommand $command
+     *
+     * @throws NoOwnerLeftException
      */
     public function handle(RemoveCollaboratorCommand $command)
     {
         $collaborator = $this->collaboratorRepository->findOneByBoardIdAndUserId($command->boardId, $command->userId);
 
+        if ($collaborator->isOwner() && $this->countByLevel($collaborator) < 2) {
+            throw new NoOwnerLeftException('You must keep at least one owner collaborator.');
+        }
+
         $this->collaboratorRepository->remove($collaborator);
+    }
+
+    /**
+     * @param $collaborator
+     *
+     * @return int
+     */
+    private function countByLevel(Collaborator $collaborator): int
+    {
+        return $this->collaboratorRepository->countByLevel($collaborator->getBoard(), $collaborator->getLevel());
     }
 }
